lifecycle-package {lifecycle}	R Documentation

lifecycle: Manage the Life Cycle of your Package Functions

Description

Manage the life cycle of your exported functions with shared conventions, documentation badges, and non-invasive deprecation warnings. The 'lifecycle' package defines four development stages (experimental, maturing, stable, and questioning) and three deprecation stages (soft-deprecated, deprecated, and defunct). It makes it easy to insert badges corresponding to these stages in your documentation. Usage of deprecated functions are signalled with increasing levels of non-invasive verbosity.

Author(s)

Maintainer: Lionel Henry lionel@rstudio.com

Other contributors:

RStudio [copyright holder]

See Also

Useful links:

https://lifecycle.r-lib.org/

https://github.com/r-lib/lifecycle

Report bugs at https://github.com/r-lib/lifecycle/issues

[Package lifecycle version 0.2.0 Index]