signal_experimental {lifecycle}	R Documentation

Signal experimental or superseded features

Description

These functions signal simple conditions with rlang::signal() (see also base::signalCondition()). They don't have any effect unless a condition handler is installed for conditions of class lifecycle_experimental or lifecycle_superseded.

Usage

signal_experimental(when, what)

signal_superseded(when, what, with = NULL)

lifecycle_cnd_data(cnd)

Arguments

when
The version at which the feature was assigned a lifecycle stage.

what, with
Feature specifications for the feature and optionally its replacement. See deprecate_soft() for details.

cnd
A condition object thrown by signal_experimental() or signal_superseded().

Details

In order to be fast, signal_experimental() and signal_superseded() do not validate their arguments. Instead, you need to catch the condition object and pass it to lifecycle_cnd_data() to get structured data.

[Package lifecycle version 0.2.0 Index]