Signal other experimental or superseded features

signal_stage() allows you to signal life cycle stages other than deprecation (for which you should use deprecate_warn() and friends). There is no behaviour associated with this signal, but in the future we will provide tools to log and report on usage of experimental and superseded functions.

signal_stage(stage, what, env = caller_env())

Arguments

stage
Life cycle stage, either "experimental" or "superseded".

what
String describing what feature the stage applies too, using the same syntax as deprecate_warn().

env
Pair of environments that define where deprecate_*() was called (used to determine the package name) and where the function called the deprecating function was called (used to determine if deprecate_soft() should message).

These are only needed if you're calling deprecate_*() from an internal helper, in which case you should forward env = caller_env() and user_env = caller_env(2).

Examples

foofy <- function(x, y, z) {
  signal_stage("experimental", "foofy()")
  x + y / z
}
foofy(1, 2, 3)

#> [1] 1.666667