pkg_lifecycle_statuses {lifecycle}	R Documentation

Lint usages of functions that have a non-stable life cycle.

Description

lint_lifecycle dynamically queries the package documentation for packages in packages for lifecycle annotations and then searches the directory in path for usages of those functions.

lint_tidyverse_lifecycle is a convenience function to call lint_lifecycle for all the packages in the tidyverse.

pkg_lifecycle_statuses returns a data frame of functions with lifecycle annotations for an installed package.

Usage

pkg_lifecycle_statuses(
  package,
  which = c("superseded", "deprecated", "questioning", "defunct", "experimental",
    "soft-deprecated", "retired")
)

lint_lifecycle(
  packages,
  path = ".",
  pattern = "[.][Rr](md)?",
  which = c("superseded", "deprecated", "questioning", "defunct", "experimental",
    "soft-deprecated", "retired")
)

lint_tidyverse_lifecycle(
  path = ".",
  pattern = "[.][Rr](md)?",
  which = c("superseded", "deprecated", "questioning", "defunct", "experimental",
    "soft-deprecated", "retired")
)

Arguments

package
The name of an installed package.

which
The lifecycle statuses to retrieve. Include NA if you want to include functions without a specified lifecycle status in the results.

packages
One or more installed packages to query for lifecycle statuses.

path
The directory path to the files you want to search.

pattern
Any files matching this pattern will be searched. The default searches any files ending in .R or .Rmd.

[Package lifecycle version 1.0.1 Index]