signal_stage {lifecycle}	R Documentation

Signal other experimental or superseded features

Description

signal_stage() allows you to signal life cycle stages other than deprecation (for which you should use deprecate_warn() and friends). There is no behaviour associated with this signal, but in the future we will provide tools to log and report on usage of experimental and superseded functions.

Usage

signal_stage(stage, what, with = NULL, env = caller_env())

Arguments

stage
Life cycle stage, either "experimental" or "superseded".

what
String describing what feature the stage applies too, using the same syntax as deprecate_warn().

with
An optional string giving a recommended replacement for a superseded function.

env
Environment used to determine where signal_stage() was called, used to determine the package name).

Examples

foofy <- function(x, y, z) {
  signal_stage("experimental", "foofy()")
  x + y / z
}
foofy(1, 2, 3)

[Package lifecycle version 1.0.3 Index]